requirements.txt
discord.py
# Google-Images-Search
Pillow
aiohttp
io